Add user signup, impl user login.
NOTE: still need to actually use this token on per-request
      authentication!!! right now, we use the dummy key...

This commit represents our front-end implementation of user
authentication. With it, a user may create an account, log in to that
account, and log out of that account. We store the user's information
locally in the browser's `localStorage` on authentication, clearing that
storage on logout.

To simplify local development, we're using `react-script`'s `NODE_ENV`
variable to determine the API's hostname (`localhost:8080` in dev,
`dev.buttaface.space` in prod) at buld time.

We add a file `Exceptions.js` that adds some custom exceptions which we
don't yet use. Once the client is more robustly built out, we should use
these instead of checking the raw HTTP 4xx/5xx error codes.

We introduce a new dependency on `jwt-decode`, a simple library to
decode valid JWT's. This commit also introduces our first use of the
browser's `history` API, which we have access to through react-router's
`props.history`.