Replace docs with tasks, add sorting and real auth.
This commit moves from the DocumentTable thingy over to a TaskTable and
add sorting functionality therein. Users can sort on the `status` and
`priority` attributes of individual tasks.

We also add support for actual authentication, sending the proper Bearer
token when available.

One thing to note is the use of React.Component's
`componentWillReceiveProps` to properly update the `data` state
attribute when TaskTableContainer passes down the task list once it
resolves its promise from the GET call to our API.

We also finally add direnv's `/.envrc` file to this repository.