Enable multi-filter by workspace.
This commit adds support to our TaskTable for filtering by the
disjunction of workspace selections. SemanticUI's `<Table/>` and its
`multiple`/`selection` props make this trivial.

We establish a new `this.state` member: `displayTasks`. This is an
ordered array of task items which we are displaying at any given
`render()`. Upon load, we initialize `displayTasks` to be the same as
`tasks`. The extant `tasks` state item is an array of all tasks
fetched from the server. Because we are only ever showing
`displayTasks`, it is sorted and filtered while `tasks` remains
effectively immutable.

To calculate the disjunction of tasks' associated workspaces and the set
of selected tasks (in order to properly update `displayTasks` on
filter), we utilize lodash's awesome `_.intersection` to detect whether
or not tasks being filtered are associated with a user-selected
workspace.

TODO
====
- make `tasks` a prop and remove it from `this.state` entirely.
- bust out all these static lists, options, etc. into a config file.
- consolidate the priority/status onChange callbacks.
- split these elements out into separate components
- add comprehensive `X` button to ws dropdown to kill all selections.

Bugs
====
- sorting with active workspace filters temporarily breaks sort
  capabilities. user needs to sort by other column and then sort by
  first column in order to remediate.